Reimplement 'happens' filter using HappensDateField
